app-info: Move XdpAppInfo into its own files
This allows us to have a clean split between utils and the XdpAppInfo
which has become a complex free standing thing.
