+++ src/network-monitor.c
@@ -25,6 +25,7 @@
 
 #include "network-monitor.h"
 #include "call.h"
+#include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
