+++ src/notification.c
@@ -27,8 +27,9 @@
 
 #include "notification.h"
 #include "call.h"
-#include "request.h"
 #include "permissions.h"
+#include "request.h"
+#include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
