+++ src/proxy-resolver.c
@@ -25,6 +25,7 @@
 
 #include "proxy-resolver.h"
 #include "call.h"
+#include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
