+++ src/realtime.c
@@ -26,6 +26,7 @@
 #include "realtime.h"
 #include "call.h"
 #include "permissions.h"
+#include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-utils.h"
