+++ src/request.h
@@ -21,6 +21,7 @@
 
 #pragma once
 
+#include "xdp-app-info.h"
 #include "xdp-utils.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
