+++ src/trash.c
@@ -35,6 +35,7 @@
 #include "trash.h"
 #include "call.h"
 #include "documents.h"
+#include "xdp-app-info.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
