+++ src/xdp-utils.h
@@ -33,13 +33,6 @@
 
 #define DESKTOP_PORTAL_OBJECT_PATH "/org/freedesktop/portal/desktop"
 
-typedef enum
-{
-  XDP_APP_INFO_KIND_HOST = 0,
-  XDP_APP_INFO_KIND_FLATPAK = 1,
-  XDP_APP_INFO_KIND_SNAP    = 2,
-} XdpAppInfoKind;
-
 gint xdp_mkstempat (int    dir_fd,
                     gchar *tmpl,
                     int    flags,
@@ -59,53 +52,10 @@ typedef void (*XdpPeerDiedCallback) (const char *name);
 typedef int XdpFd;
 G_DEFINE_AUTO_CLEANUP_FREE_FUNC(XdpFd, close, -1)
 
-#define XDP_TYPE_APP_INFO (xdp_app_info_get_type())
-G_DECLARE_FINAL_TYPE (XdpAppInfo,
-                      xdp_app_info,
-                      XDP, APP_INFO,
-                      GObject)
-
-const char *xdp_app_info_get_id          (XdpAppInfo  *app_info);
-char *      xdp_app_info_get_instance    (XdpAppInfo  *app_info);
-gboolean    xdp_app_info_is_host         (XdpAppInfo  *app_info);
-gboolean    xdp_app_info_is_flatpak      (XdpAppInfo  *app_info);
-XdpAppInfoKind xdp_app_info_get_kind     (XdpAppInfo  *app_info);
-char *      xdp_app_info_remap_path      (XdpAppInfo  *app_info,
-                                          const char  *path);
-gboolean    xdp_app_info_map_pids        (XdpAppInfo  *app_info,
-                                          pid_t       *pids,
-                                          guint        n_pids,
-                                          GError     **error);
-gboolean    xdp_app_info_map_tids        (XdpAppInfo  *app_info,
-                                          pid_t        owner_pid,
-                                          pid_t       *tids,
-                                          guint        n_tids,
-                                          GError     **error);
-gboolean    xdp_app_info_pidfds_to_pids (XdpAppInfo  *app_info,
-                                         const int   *fds,
-                                         pid_t       *pids,
-                                         gint         count,
-                                         GError     **error);
-char *      xdp_app_info_get_path_for_fd (XdpAppInfo  *app_info,
-                                          int          fd,
-                                          int          require_st_mode,
-                                          struct stat *st_buf,
-                                          gboolean    *writable_out,
-                                          GError     **error);
-gboolean    xdp_app_info_has_network     (XdpAppInfo  *app_info);
-GAppInfo *  xdp_app_info_load_app_info   (XdpAppInfo *app_info);
-char **     xdp_app_info_rewrite_commandline (XdpAppInfo        *app_info,
-                                              const char *const *commandline,
-                                              gboolean           quote_escape);
-char       *xdp_app_info_get_tryexec_path (XdpAppInfo  *app_info);
-
 void xdp_set_documents_mountpoint (const char *path);
 const char * xdp_get_documents_mountpoint (void);
 char * xdp_get_alternate_document_path (const char *path, const char *app_id);
 
-XdpAppInfo *xdp_invocation_lookup_app_info_sync (GDBusMethodInvocation *invocation,
-                                                 GCancellable          *cancellable,
-                                                 GError               **error);
 void   xdp_connection_track_name_owners  (GDBusConnection       *connection,
                                           XdpPeerDiedCallback    peer_died_cb);
 
@@ -187,10 +137,3 @@ gboolean xdp_spawnv     (GFile                *dir,
 char * xdp_canonicalize_filename (const char *path);
 gboolean  xdp_has_path_prefix (const char *str,
                                const char *prefix);
-
-/* exposed for the benefit of tests */
-int _xdp_parse_cgroup_file (FILE     *f,
-                            gboolean *is_snap);
-#ifdef HAVE_LIBSYSTEMD
-char *_xdp_parse_app_id_from_unit_name (const char *unit);
-#endif
